StudyMake DB와 연결 APIS에 post 함수 추가 PostDataModel.kt에 PostContentUploadDTO 데이터 타입 생성 StudyMakeNotification.kt에서 Studymake에 관한 모든 데이터들 post

+++ app/src/main/java/com/itstime/haejo/study/StudyMakeActivity.kt
@@ -107,7 +107,7 @@ class StudyMakeActivity : AppCompatActivity() {
             res.getStringArray(R.array.week)[binding.spinnerWeek.selectedItemPosition],
             res.getStringArray(R.array.isUntact)[binding.spinnerisUntact.selectedItemPosition],
             res.getStringArray(R.array.category)[binding.spinnerCategory.selectedItemPosition],
-            res.getStringArray(R.array.peopleAmount)[binding.spinnerPeopleAmount.selectedItemPosition][0] - '0',
+            (res.getStringArray(R.array.peopleAmount)[binding.spinnerPeopleAmount.selectedItemPosition][0] - '0').toInt(),
             binding.etSurveyContent.text.toString(),
             null,
             null,
